refactor(multisig): move hash from types to exported
